requirements.txt
tabulate==0.10.0
python-dotenv==1.2.2
